% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pclExport.r
\name{pclExport}
\alias{pclExport}
\title{Test read & export}
\usage{
pclExport(mesh)
}
\arguments{
\item{mesh}{mesh to read}
}
\value{
\item{vb}{original vertices of the mesh }
}
\description{
calculate common allometric component
}

