% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pclRegister.r
\name{pclRegister}
\alias{pclRegister}
\title{Pose estimation through spine images}
\usage{
pclRegister(meshRef, meshTar, radiusSearch = 100, imageWidth = 8,
  maxIt = 500, numberOfSample = 3, correspondenceRandomness = 5,
  similarityThreshold = 0.9, inlierFraction = 0.25)
}
\arguments{
\item{meshRef}{triangular mesh of class "mesh3d": a reference mesh.}

\item{meshTar}{triangular mesh of class "mesh3d": a target mesh.}

\item{radiusSearch}{numeric: radius of the support cylinder.}

\item{imageWidth}{unsigned int: resolution of the spin image (the number of bins along one dimension), cannot be changed yet.}

\item{maxIt}{int: number of RANSAC iterations.}

\item{numberOfSample}{int: number of points to sample for generating/prerejecting a pose.}

\item{correspondenceRandomness}{int: number of nearest features to use.}

\item{similarityThreshold}{numeric: polygonal edge length similarity threshold.}

\item{inlierFraction}{numeric: required inlier fraction for accepting a pose hypothesis.}
}
\value{
the registered reference mesh
}
\description{
Pose estimation through spine images
}
\author{
François Girinon
}

