% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rpcl-package.R
\docType{package}
\name{Rpcl-package}
\alias{Rpcl}
\alias{Rpcl-package}
\title{Integrating  PCL and R}
\description{
Integrating PCL and R
}
\details{
\tabular{ll}{
Package: \tab PCL\cr
Type: \tab Package\cr
Version: \tab 0.2.150716\cr
Date: \tab 2015-07-16\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\author{
François Girinon

Maintainer: François Girinon <francois.girinon@gadz.org>
}
\references{
To be announced
}
\keyword{package}

